equip weapons from pickup
- added a singleton for weapon pickups types
- moved code to pickup to player script
- TODO: https://docs.godotengine.org/en/stable/tutorials/scripting/singletons_autoload.html
use this way for c# singletons
