+++ project.godot
@@ -15,6 +15,10 @@ run/main_scene="res://Levels/Scenes/level_one.tscn"
 config/features=PackedStringArray("4.3", "C#", "Mobile")
 config/icon="res://icon.svg"
 
+[autoload]
+
+WeaponTypes="*res://Singletons/WeaponTypes.cs"
+
 [display]
 
 window/size/resizable=false
